py/requirements.txt
RPi.GPIO==0.7.0
bottle==0.12.15
